requirements.txt
# --- Core Deep Learning ---
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0

# --- Federated Learning Framework ---
flwr==1.7.0
flwr-datasets==0.0.2

# --- Visualization ---
matplotlib==3.8.2
seaborn==0.13.0

# --- Multi-modal Processing ---
pillow==10.1.0
opencv-python==4.8.1.78
sentence-transformers==2.2.2

# --- Distributed Computing ---
dask[complete]==2023.12.0

# --- API & Dashboard ---
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.29.0
plotly==5.18.0
python-multipart==0.0.6
requests==2.31.0

# --- Data Processing ---
pandas==2.1.3
numpy==1.26.3
scikit-learn==1.3.2

# --- Privacy ---
opacus==1.4.0

# --- Utilities ---
pyyaml==6.0.1
tqdm==4.66.1
loguru==0.7.2
python-dotenv==1.0.0

# --- Development ---
jupyter==1.0.0
ipykernel==6.27.1
pytest==7.4.3
black==23.12.0

# --- Vector Database  ---
pymilvus==2.4.0
marshmallow==3.21.0

httpx==0.25.2